creates project-specific subfolders to make file structure easier to read
