+++ src/App.tsx
@@ -23,23 +23,23 @@ import { Route, BrowserRouter as Router, Switch } from "react-router-dom";
 import About from "./components/About";
 import Advertising from "./components/Advertising";
 import Background from "./components/Background";
-import Belledejour from "./components/projects/Belledejour";
-import ConsciousShopping from "./components/projects/ConsciousShopping";
+import Belledejour from "./components/projects/belledejour/Belledejour";
+import ConsciousShopping from "./components/projects/conscious-shopping/ConsciousShopping";
 import Contact from "./components/Contact";
-import EroticStories from "./components/projects/EroticStories";
+import EroticStories from "./components/projects/erotic-stories/EroticStories";
 import Error404 from "./components/Error404";
-import Eye from "./components/projects/Eye";
-import FashionEditorial from "./components/projects/FashionEditorial";
+import Eye from "./components/projects/eye/Eye";
+import FashionEditorial from "./components/projects/fashion-editorial/FashionEditorial";
 import Flex from "./components/Flex";
 import GlobalStyle from "./GlobalStyle";
 import Header from "./components/Header";
 import Home from "./components/Home";
-import KaiLandre from "./components/projects/KaiLandre";
-import LeoAdef from "./components/projects/LeoAdef";
+import KaiLandre from "./components/projects/kai-landre/KaiLandre";
+import LeoAdef from "./components/projects/leo-adef/LeoAdef";
 import Loading from "./components/Loading";
 import Manifesto from "./components/Manifesto";
-import Map from "./components/projects/Map";
-import MarcMedina from "./components/projects/MarcMedina";
+import Map from "./components/projects/map/Map";
+import MarcMedina from "./components/projects/marc-medina/MarcMedina";
 import Projects from "./components/Projects";
 import { ThemeProvider } from "styled-components";
 import theme from "./components/theme";
