+++ src/components/projects/belledejour/Belledejour.tsx
@@ -1,9 +1,9 @@
 import { TypographyProps, typography } from "styled-system";
 
-import EroticStoriesContent from "./EroticStoriesContent";
-import EroticStoriesPreview from "./EroticStoriesPreview";
-import Flex from "../Flex";
-import PreviewOrProjectPage from "./PreviewOrProjectPage";
+import BelledejourContent from "./BelledejourContent";
+import BelledejourPreview from "./BelledejourPreview";
+import Flex from "../../Flex";
+import PreviewOrProjectPage from "../PreviewOrProjectPage";
 import React from "react";
 import styled from "styled-components";
 
@@ -12,7 +12,7 @@ const Main = styled(Flex)<TypographyProps>`
 `;
 
 const launchDate = "2021-03-21";
-const EroticStories = () => {
+const Belledejour = () => {
   return (
     <Main
       flex="auto"
@@ -22,11 +22,11 @@ const EroticStories = () => {
     >
       <PreviewOrProjectPage
         launchDate={launchDate}
-        PreviewPage={EroticStoriesPreview}
-        ProjectPage={EroticStoriesContent}
+        PreviewPage={BelledejourPreview}
+        ProjectPage={BelledejourContent}
       />
     </Main>
   );
 };
 
-export default EroticStories;
+export default Belledejour;
