+++ src/components/projects/belledejour/BelledejourContent.tsx
@@ -1,23 +1,23 @@
 import { LayoutProps, layout } from "styled-system";
 
-import Grid from "../Grid";
+import Grid from "../../Grid";
 import React from "react";
-import image1 from "../../assets/project-page/belledejour/belledejour-1.jpg";
-import image10 from "../../assets/project-page/belledejour/belledejour-10.jpg";
-import image11 from "../../assets/project-page/belledejour/belledejour-11.jpg";
-import image12 from "../../assets/project-page/belledejour/belledejour-12.jpg";
-import image13 from "../../assets/project-page/belledejour/belledejour-13.jpg";
-import image14 from "../../assets/project-page/belledejour/belledejour-14.jpg";
-import image15 from "../../assets/project-page/belledejour/belledejour-15.jpg";
-import image16 from "../../assets/project-page/belledejour/belledejour-16.jpg";
-import image2 from "../../assets/project-page/belledejour/belledejour-2.jpg";
-import image3 from "../../assets/project-page/belledejour/belledejour-3.jpg";
-import image4 from "../../assets/project-page/belledejour/belledejour-4.jpg";
-import image5 from "../../assets/project-page/belledejour/belledejour-5.jpg";
-import image6 from "../../assets/project-page/belledejour/belledejour-6.jpg";
-import image7 from "../../assets/project-page/belledejour/belledejour-7.jpg";
-import image8 from "../../assets/project-page/belledejour/belledejour-8.jpg";
-import image9 from "../../assets/project-page/belledejour/belledejour-9.jpg";
+import image1 from "../../../assets/project-page/belledejour/belledejour-1.jpg";
+import image10 from "../../../assets/project-page/belledejour/belledejour-10.jpg";
+import image11 from "../../../assets/project-page/belledejour/belledejour-11.jpg";
+import image12 from "../../../assets/project-page/belledejour/belledejour-12.jpg";
+import image13 from "../../../assets/project-page/belledejour/belledejour-13.jpg";
+import image14 from "../../../assets/project-page/belledejour/belledejour-14.jpg";
+import image15 from "../../../assets/project-page/belledejour/belledejour-15.jpg";
+import image16 from "../../../assets/project-page/belledejour/belledejour-16.jpg";
+import image2 from "../../../assets/project-page/belledejour/belledejour-2.jpg";
+import image3 from "../../../assets/project-page/belledejour/belledejour-3.jpg";
+import image4 from "../../../assets/project-page/belledejour/belledejour-4.jpg";
+import image5 from "../../../assets/project-page/belledejour/belledejour-5.jpg";
+import image6 from "../../../assets/project-page/belledejour/belledejour-6.jpg";
+import image7 from "../../../assets/project-page/belledejour/belledejour-7.jpg";
+import image8 from "../../../assets/project-page/belledejour/belledejour-8.jpg";
+import image9 from "../../../assets/project-page/belledejour/belledejour-9.jpg";
 import styled from "styled-components";
 
 const BelledejourImageContainer = styled.div<LayoutProps>`
