+++ src/components/projects/erotic-stories/EroticStoriesPreview.tsx
@@ -1,18 +1,20 @@
 import { LayoutProps, layout } from "styled-system";
 
-import Flex from "../Flex";
+import Flex from "../../Flex";
 import { Link } from "react-router-dom";
-import { PROJECTS_URL } from "../../constants/router-urls";
+import { PROJECTS_URL } from "../../../constants/router-urls";
 import React from "react";
-import Timer from "../Timer";
-import stairs from "../../assets/project-page/project-icons/stairs.png";
+import Timer from "../../Timer";
+import statue from "../../../assets/project-page/project-icons/statue.png";
 import styled from "styled-components";
 
 const Img = styled.img<LayoutProps>`
   ${layout};
 `;
 
-const FashionEditorial: React.FC<{ launchDate: string }> = ({ launchDate }) => {
+const EroticStoriesPreview: React.FC<{ launchDate: string }> = ({
+  launchDate,
+}) => {
   return (
     <Flex
       flex="auto"
@@ -23,7 +25,7 @@ const FashionEditorial: React.FC<{ launchDate: string }> = ({ launchDate }) => {
     >
       <Link to={PROJECTS_URL}>
         <Flex justifyContent="center" alignItems="center">
-          <Img src={stairs} alt="stairs icon" maxWidth="30%" />
+          <Img src={statue} alt="statue icon" maxWidth="30%" />
         </Flex>
       </Link>
       <Timer launchDate={launchDate} />
@@ -31,4 +33,4 @@ const FashionEditorial: React.FC<{ launchDate: string }> = ({ launchDate }) => {
   );
 };
 
-export default FashionEditorial;
+export default EroticStoriesPreview;
