+++ src/components/projects/eye/EyeContent.tsx
@@ -10,16 +10,16 @@ import {
   layout,
   typography,
 } from "styled-system";
-import { DataSet, DataValues, EyeData } from "../../fixtures/eye-data";
+import { DataSet, DataValues, EyeData } from "../../../fixtures/eye-data";
 import React, { useCallback, useState } from "react";
 import styled, { keyframes } from "styled-components";
 
-import Flex from "../Flex";
-import Grid from "../Grid";
+import Flex from "../../Flex";
+import Grid from "../../Grid";
 import { Picker } from "react-mobile-style-picker";
-import data from "../../fixtures/eye-data";
-import eyeBlue from "../../assets/project-page/eye/eye_blue.png";
-import eyeBrown from "../../assets/project-page/eye/eye_brown.png";
+import data from "../../../fixtures/eye-data";
+import eyeBlue from "../../../assets/project-page/eye/eye_blue.png";
+import eyeBrown from "../../../assets/project-page/eye/eye_brown.png";
 
 const P = styled.p<BorderProps & TypographyProps>`
   text-transform: uppercase;
