+++ src/components/projects/fashion-editorial/FashionEditorialContent.tsx
@@ -1,30 +1,30 @@
 import { LayoutProps, layout } from "styled-system";
 
-import Grid from "../Grid";
+import Grid from "../../Grid";
 import React from "react";
-import image1 from "../../assets/project-page/fashion-editorial/awewave-oxymore-01.jpg";
-import image10 from "../../assets/project-page/fashion-editorial/awewave-oxymore-10.jpg";
-import image11 from "../../assets/project-page/fashion-editorial/awewave-oxymore-11.jpg";
-import image12 from "../../assets/project-page/fashion-editorial/awewave-oxymore-12.jpg";
-import image13 from "../../assets/project-page/fashion-editorial/awewave-oxymore-13.jpg";
-import image14 from "../../assets/project-page/fashion-editorial/awewave-oxymore-14.jpg";
-import image15 from "../../assets/project-page/fashion-editorial/awewave-oxymore-15.jpg";
-import image16 from "../../assets/project-page/fashion-editorial/awewave-oxymore-16.jpg";
-import image17 from "../../assets/project-page/fashion-editorial/awewave-oxymore-17.jpg";
-import image18 from "../../assets/project-page/fashion-editorial/awewave-oxymore-18.jpg";
-import image19 from "../../assets/project-page/fashion-editorial/awewave-oxymore-19.jpg";
-import image2 from "../../assets/project-page/fashion-editorial/awewave-oxymore-02.jpg";
-import image20 from "../../assets/project-page/fashion-editorial/awewave-oxymore-20.jpg";
-import image21 from "../../assets/project-page/fashion-editorial/awewave-oxymore-21.jpg";
-import image22 from "../../assets/project-page/fashion-editorial/amewave-oxymore-22.jpg";
-import image23 from "../../assets/project-page/fashion-editorial/amewave-oxymore-23.jpg";
-import image3 from "../../assets/project-page/fashion-editorial/awewave-oxymore-03.jpg";
-import image4 from "../../assets/project-page/fashion-editorial/awewave-oxymore-04.jpg";
-import image5 from "../../assets/project-page/fashion-editorial/awewave-oxymore-05.jpg";
-import image6 from "../../assets/project-page/fashion-editorial/awewave-oxymore-06.jpg";
-import image7 from "../../assets/project-page/fashion-editorial/awewave-oxymore-07.jpg";
-import image8 from "../../assets/project-page/fashion-editorial/awewave-oxymore-08.jpg";
-import image9 from "../../assets/project-page/fashion-editorial/awewave-oxymore-09.jpg";
+import image1 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-01.jpg";
+import image10 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-10.jpg";
+import image11 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-11.jpg";
+import image12 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-12.jpg";
+import image13 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-13.jpg";
+import image14 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-14.jpg";
+import image15 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-15.jpg";
+import image16 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-16.jpg";
+import image17 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-17.jpg";
+import image18 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-18.jpg";
+import image19 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-19.jpg";
+import image2 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-02.jpg";
+import image20 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-20.jpg";
+import image21 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-21.jpg";
+import image22 from "../../../assets/project-page/fashion-editorial/amewave-oxymore-22.jpg";
+import image23 from "../../../assets/project-page/fashion-editorial/amewave-oxymore-23.jpg";
+import image3 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-03.jpg";
+import image4 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-04.jpg";
+import image5 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-05.jpg";
+import image6 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-06.jpg";
+import image7 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-07.jpg";
+import image8 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-08.jpg";
+import image9 from "../../../assets/project-page/fashion-editorial/awewave-oxymore-09.jpg";
 import styled from "styled-components";
 
 const EditorialImageContainer = styled.div<LayoutProps>`
