+++ src/components/projects/fashion-editorial/FashionEditorialPreview.tsx
@@ -1,21 +1,18 @@
 import { LayoutProps, layout } from "styled-system";
 
-import Flex from "../Flex";
+import Flex from "../../Flex";
 import { Link } from "react-router-dom";
-import { PROJECTS_URL } from "../../constants/router-urls";
+import { PROJECTS_URL } from "../../../constants/router-urls";
 import React from "react";
-import Timer from "../Timer";
-import dragon from "../../assets/project-page/project-icons/dragon.png";
+import Timer from "../../Timer";
+import stairs from "../../../assets/project-page/project-icons/stairs.png";
 import styled from "styled-components";
 
 const Img = styled.img<LayoutProps>`
   ${layout};
 `;
 
-interface KaiLandrePreviewProps {
-  launchDate: string;
-}
-const KaiLandrePreview: React.FC<KaiLandrePreviewProps> = ({ launchDate }) => {
+const FashionEditorial: React.FC<{ launchDate: string }> = ({ launchDate }) => {
   return (
     <Flex
       flex="auto"
@@ -26,7 +23,7 @@ const KaiLandrePreview: React.FC<KaiLandrePreviewProps> = ({ launchDate }) => {
     >
       <Link to={PROJECTS_URL}>
         <Flex justifyContent="center" alignItems="center">
-          <Img src={dragon} alt="dragon icon" maxWidth="30%" />
+          <Img src={stairs} alt="stairs icon" maxWidth="30%" />
         </Flex>
       </Link>
       <Timer launchDate={launchDate} />
@@ -34,4 +31,4 @@ const KaiLandrePreview: React.FC<KaiLandrePreviewProps> = ({ launchDate }) => {
   );
 };
 
-export default KaiLandrePreview;
+export default FashionEditorial;
