+++ src/components/projects/map/Map.tsx
@@ -1,13 +1,13 @@
 import { LayoutProps, layout } from "styled-system";
 import React, { Fragment } from "react";
 
-import Flex from "../Flex";
+import Flex from "../../Flex";
 import { Link } from "react-router-dom";
 import MapContent from "./MapContent";
-import { PROJECTS_URL } from "../../constants/router-urls";
-import PreviewOrProjectPage from "./PreviewOrProjectPage";
-import Timer from "../Timer";
-import magnify from "../../assets/project-page/project-icons/magnify.png";
+import { PROJECTS_URL } from "../../../constants/router-urls";
+import PreviewOrProjectPage from "../PreviewOrProjectPage";
+import Timer from "../../Timer";
+import magnify from "../../../assets/project-page/project-icons/magnify.png";
 import styled from "styled-components";
 
 const Img = styled.img<LayoutProps>`
